README.md
This is an application built to give the latest reports of earthquakes taking place throughout the world. This is an android app that uses the usgs api for gathering the json format data. It gives the magnitude(With color, showing how dangerous it is.), the place where it occured and date. By clicking on any if the reports, it directly takes you to the USGS site for the full detailed information about the earthquake. Also there is a setting for minimum magnitude.